Here’s a high-level overview of what it does:
It first cleans up any old token and network ID files that are more than an hour old.
It retrieves an access token, either from a stored file or by making a request to get a new one.
It fetches the list of available network names.
It prompts the user to enter the name of the network they want to interact with.
It retrieves the ID of the specified network.
It prompts the user to specify the number of voucher groups they want to create and collects necessary details for each group.
It creates the specified number of voucher groups.
It retrieves the list of all voucher groups.
For each voucher group, it retrieves the list of vouchers and appends them to a list.
Finally, it returns the list of all vouchers.
This function allows you to interact with a specified network, create voucher groups, and retrieve information about vouchers within those groups. The returned list of vouchers can then be used for further operations as needed. Please note that this function involves user interaction for inputting network and voucher details.

The function get_voucher_group_list is used to get the list of voucher groups in a network. It makes a POST request to the /oapi/v1.0.0/voucher/list endpoint with the necessary parameters and returns the IDs of the voucher groups.

The function get_voucher_list_in_group is used to get the list of vouchers in a specific voucher group. It makes a POST request to the /oapi/v1.0.0/voucher/vouchers/list endpoint with the necessary parameters and returns the IDs of the vouchers in the group.

These functions are part of a larger process of interacting with the network, creating voucher groups, and retrieving voucher information. They are used within the interact_with_network function.